isolate definitions in their own cpp
